requirements.txt
fastapi
uvicorn[standard]
stripe
python-multipart
pydantic
python-dotenv
reportlab
requests 
pydantic[email]
pandas
jinja2
aiofiles